Renamed tests and suites
