+++ src/main/groovy/liquibase/harness/BaseHarnessSuite.groovy
@@ -7,6 +7,6 @@ import org.junit.runner.RunWith
 import org.junit.runners.Suite
 
 @RunWith(Suite.class)
-@Suite.SuiteClasses([ChangeObjectTests, ChangeDataTests, SnapshotObjectTests])
+@Suite.SuiteClasses([ChangeObjectTests, ChangeDataTests])
 abstract class BaseHarnessSuite {
 }
