+++ src/main/groovy/liquibase/harness/FoundationalHarnessSuite.groovy
@@ -1,12 +1,11 @@
 package liquibase.harness
 
 
-import liquibase.harness.compatibility.basic.CompatibilityTest
 import liquibase.harness.compatibility.foundational.FoundationalTest
 import org.junit.runner.RunWith
 import org.junit.runners.Suite
 
 @RunWith(Suite.class)
-@Suite.SuiteClasses([CompatibilityTest, FoundationalTest])
+@Suite.SuiteClasses([FoundationalTest])
 abstract class FoundationalHarnessSuite {
 }
