+++ src/main/groovy/liquibase/harness/compatibility/foundational/FoundationalTestHelper.groovy
@@ -4,72 +4,70 @@ import groovy.transform.ToString
 import groovy.transform.builder.Builder
 import liquibase.Scope
 import liquibase.database.Database
+import liquibase.database.DatabaseConnection
+import liquibase.database.jvm.JdbcConnection
 import liquibase.harness.config.DatabaseUnderTest
 import liquibase.harness.config.TestConfig
 import liquibase.harness.util.DatabaseConnectionUtil
 import liquibase.harness.util.FileUtils
-import liquibase.ui.UIService
+import java.sql.Connection
+import java.sql.DriverManager
+import java.sql.ResultSet
+import java.sql.SQLException
 
 class FoundationalTestHelper {
-    final static String baseChangelogPath = "liquibase/harness/compatibility/foundational/"
-    final static UIService uiService = Scope.getCurrentScope().getUI()
+
+    final static String baseChangelogPath = "liquibase/harness/compatibility/basic/changelogs"
+    final static List supportedChangeLogFormats = ['xml', 'sql', 'json', 'yml', 'yaml'].asImmutable()
+
+    static boolean shouldOpenNewConnection(DatabaseConnection connection, String... dbNames) {
+        return connection.isClosed()||Arrays.stream(dbNames).anyMatch({ dbName -> connection.getDatabaseProductName().toLowerCase().contains(dbName) })
+    }
 
     static List<TestInput> buildTestInput() {
-        String commandLineChanges = System.getProperty("change")
-        List commandLineChangesList = Collections.emptyList()
-        if (commandLineChanges) {
-            commandLineChangesList = Arrays.asList(commandLineChanges.contains(",")
-                    ? commandLineChanges.split(",")
-                    : commandLineChanges)
+        String commandLineInputFormat = System.getProperty("inputFormat")
+        if (commandLineInputFormat) {
+            if (!supportedChangeLogFormats.contains(commandLineInputFormat)) {
+                throw new IllegalArgumentException(commandLineInputFormat + " inputFormat is not supported")
+            }
+            TestConfig.instance.inputFormat = commandLineInputFormat
         }
+        Scope.getCurrentScope().getUI().sendMessage("Only " + TestConfig.instance.inputFormat
+                + " input files are taken into account for this test run")
 
         List<TestInput> inputList = new ArrayList<>()
         DatabaseConnectionUtil databaseConnectionUtil = new DatabaseConnectionUtil()
         for (DatabaseUnderTest databaseUnderTest : databaseConnectionUtil
                 .initializeDatabasesConnection(TestConfig.instance.getFilteredDatabasesUnderTest())) {
-            for (def changeLogEntry : FileUtils.resolveInputFilePaths(databaseUnderTest, baseChangelogPath +
-                    "expectedChangeLog", "sql").entrySet()) {
-                if (!commandLineChangesList || commandLineChangesList.contains(changeLogEntry.key)) {
-                    inputList.add(TestInput.builder()
-                            .databaseName(databaseUnderTest.name)
-                            .url(databaseUnderTest.url)
-                            .dbSchema(databaseUnderTest.dbSchema)
-                            .username(databaseUnderTest.username)
-                            .password(databaseUnderTest.password)
-                            .version(databaseUnderTest.version)
-                            .setupChangelogPath(changeLogEntry.value)
-                            .insertChangelogPath(FileUtils.resolveInputFilePaths(databaseUnderTest, baseChangelogPath +
-                                    "stress/insert", "xml").get(changeLogEntry.key))
-                            .updateChangelogPath(FileUtils.resolveInputFilePaths(databaseUnderTest, baseChangelogPath +
-                                    "stress/update", "xml").get(changeLogEntry.key))
-                            .selectChangelogPath(FileUtils.resolveInputFilePaths(databaseUnderTest, baseChangelogPath +
-                                    "stress/select", "xml").get(changeLogEntry.key))
-                            .xmlChangelogPath(FileUtils.resolveInputFilePaths(databaseUnderTest, baseChangelogPath +
-                                    "expectedChangeLog", "xml").get(changeLogEntry.key))
-                            .sqlChangelogPath(FileUtils.resolveInputFilePaths(databaseUnderTest, baseChangelogPath +
-                                    "expectedChangeLog", "sql").get(changeLogEntry.key))
-                            .change(changeLogEntry.key)
-                            .database(databaseUnderTest.database)
-                            .build())
-                }
+            for (def changeLogEntry : FileUtils.resolveInputFilePaths(databaseUnderTest, baseChangelogPath, "xml").entrySet()) {
+                inputList.add(TestInput.builder()
+                        .databaseName(databaseUnderTest.name)
+                        .url(databaseUnderTest.url)
+                        .dbSchema(databaseUnderTest.dbSchema)
+                        .username(databaseUnderTest.username)
+                        .password(databaseUnderTest.password)
+                        .version(databaseUnderTest.version)
+                        .change(changeLogEntry.key)
+                        .database(databaseUnderTest.database)
+                        .build())
             }
         }
         return inputList
     }
 
-    static validateSqlChangelog(String expectedSqlChangelog, String generatedSqlChangelog) {
-        String replacementRegexp = "--(.*?)\r?\n" //removes all sql comments starting from "--" till the end of line
-        String replacementRegexpNoEOL = "--(.*?)\$" //removes all sql comments starting from "--" till the end of file
-        String cleanExpectedChangelog = expectedSqlChangelog
-                .replaceAll(replacementRegexp, "")
-                .replaceAll(replacementRegexpNoEOL, "")
-                .trim()
-        String cleanGeneratedChangelog = generatedSqlChangelog
-                .replaceAll(replacementRegexp, "")
-                .replaceAll(replacementRegexpNoEOL, "")
-                .trim()
-        assert cleanExpectedChangelog.equalsIgnoreCase(cleanGeneratedChangelog)
-        uiService.sendMessage("GENERATED SQL CHANGELOG: \n $cleanGeneratedChangelog \n EXPECTED SQL CHANGELOG: \n $cleanExpectedChangelog")
+    static ResultSet executeQuery(String pathToSql, TestInput testInput) throws SQLException {
+        Connection newConnection
+        ResultSet resultSet
+        if (shouldOpenNewConnection(testInput.database.getConnection(), "firebird")) {
+            newConnection = DriverManager.getConnection(testInput.url, testInput.username, testInput.password)
+            resultSet = newConnection.createStatement().executeQuery(pathToSql)
+            newConnection.close()
+        } else {
+            JdbcConnection connection = (JdbcConnection) testInput.database.connection
+            resultSet = connection.createStatement().executeQuery(pathToSql)
+            testInput.database.connection.autoCommit ?: testInput.database.connection.commit()
+        }
+        return resultSet
     }
 
     @Builder
@@ -80,12 +78,6 @@ class FoundationalTestHelper {
         String username
         String password
         String url
-        String setupChangelogPath
-        String insertChangelogPath
-        String updateChangelogPath
-        String selectChangelogPath
-        String xmlChangelogPath
-        String sqlChangelogPath
         String dbSchema
         String change
         Database database
